fix(desktop-notifications): session gets properly compared again
This bug was introduced by commit https://github.com/Icinga/icinga-notifications-web/pull/146/commits/6f1ec2d26f54e24dc9d15cdeab4c634ed6a1fb7a as it started treating the cookies array as a nested array, which it is not. It thus compared the first character of the submitted session identifier against the full identifier in the database.