fix(Connection): connection gets properly mapped for both IPv4 and IPv6 socket bindings
This fixes an issue which caused the URI mapping of the connections to fail when not using the IPv6 notation for both IPv4 and IPv6 addresses. It now supports both formats (IPv4 and IPv6 notation) and additionally resolves IPv4 addresses that are passed in the IPv6 notation.
Tested with: native IPv4, native IPv4 in IPv6 notation, native IPv6