Handle unicode in contact and contact group names
Accessing $string[0] returns the first byte, which may be only part of a UTF-8
encoded character. In that case, that byte on its own is invalid UTF-8 and will
be replaced with the unicode replacement character which is then shown as the
contact or contact group icon.

This commit fixes it by using grapheme_substr() instead which handles this
properly. Note that mb_substr() also exists, but that wouldn't handle even more
complex characters composed of multiple code points (like some emojis)
correctly.