EventRulesController: Do not update col1 upon adding a rule
That's a remnant of a previous implementation. It causes
two parallel requests, both access and update the session.
The last request updating the session overrides the changes
of the other and in case of xDebug being disabled or inactive
the timing of the requests is different, hence I did not
notice this prior. So the actual problem is not solved…

fixes #355