contact-groups: Prevent new groups without a channel
Without channels, most of the configuration is not possible.
Technically, channels are not required for groups, but we
want to guide new users to the channel configuration because
of its importance. Once there is a channel, groups can be
created but we still hint that there are no contacts yet.