Use past time since the day start to display clock
Use the current timestamp rounded to the nearest thirty minutes to
take a diff to the day's start at 12 am. Use the past hours * 2
plus 1 if the half hour is past to align the clock correctly. If
the rounded current time is 12 am at the next day the diff will
return **0 past hours**, but **1 past day**. For this case we will
multiply by **24**.