Use session for display timezone
Instead of using an url param that has to be preserved by every
link, form action or redirect, we now use the 'notifications'
session namespace to store the display timezone for the current
user. The start day for the timeline now comes from the
controller, no longer from the schedule detail controls.