Timeline: Make sure the header overlaps everything again
Didn't succeed while trying to fix the flyout positioning.
The problem is that since the header is a sticky element,
it's not influencing the containers offsetHeight/scrollHeight
and so there's no way to know for sure where the constraints
of the box really are. But the workaround for the user should
be straightforward anyway: Just scroll the content down again.

fixes #399